Move token source logic into own file and add more tests
To make it easier to find the logic and to make sure that it works as
intended.
