+++ config/new_test.go
@@ -5,11 +5,8 @@ package config
 
 import (
 	"crypto/tls"
-	"io/ioutil"
 	"testing"
 
-	"github.com/hashicorp/hcp-sdk-go/auth"
-	"github.com/hashicorp/hcp-sdk-go/auth/workload"
 	"github.com/hashicorp/hcp-sdk-go/profile"
 	requirepkg "github.com/stretchr/testify/require"
 )
@@ -151,71 +148,4 @@ func TestNew_NoConfigPassed(t *testing.T) {
 	require.Equal(defaultPortalURL, config.PortalURL().String())
 	require.Equal(defaultAPIAddress, config.APIAddress())
 	require.Equal(defaultSCADAAddress, config.SCADAAddress())
-}
-
-func TestNew_CredentialFile_Workload(t *testing.T) {
-	require := requirepkg.New(t)
-
-	// Write the cred file
-	cf := &auth.CredentialFile{
-		ProjectID: "123",
-		Scheme:    auth.CredentialFileSchemeWorkload,
-		Workload: &workload.IdentityProviderConfig{
-			ProviderResourceName: "iam/test",
-			AWS:                  &workload.AWSCredentialSource{},
-		},
-	}
-
-	f, err := ioutil.TempFile("", "")
-	require.NoError(err)
-	require.NoError(auth.WriteCredentialFile(f.Name(), cf))
-
-	// Exercise
-	t.Setenv(auth.EnvHCPCredFile, f.Name())
-	config, err := NewHCPConfig()
-	require.NoError(err)
-
-	rawConfig, ok := config.(*hcpConfig)
-	require.True(ok)
-
-	tokenSource, tokenSourceType, sourceIdentifier, err := rawConfig.getTokenSource()
-	require.NoError(err)
-
-	_, ok = tokenSource.(*workload.Provider)
-	require.True(ok)
-	require.Equal(sourceTypeWorkload, tokenSourceType)
-	require.Equal("iam/test", sourceIdentifier)
-}
-
-func TestNew_CredentialFile_ServicePrincipal(t *testing.T) {
-	require := requirepkg.New(t)
-
-	// Write the cred file
-	cf := &auth.CredentialFile{
-		ProjectID: "123",
-		Scheme:    auth.CredentialFileSchemeServicePrincipal,
-		Oauth: &auth.OauthConfig{
-			ClientID:     "client-id",
-			ClientSecret: "client-secret",
-		},
-	}
-
-	f, err := ioutil.TempFile("", "")
-	require.NoError(err)
-	require.NoError(auth.WriteCredentialFile(f.Name(), cf))
-
-	// Exercise
-	t.Setenv(auth.EnvHCPCredFile, f.Name())
-	config, err := NewHCPConfig()
-	require.NoError(err)
-
-	rawConfig, ok := config.(*hcpConfig)
-	require.True(ok)
-
-	tokenSource, tokenSourceType, sourceIdentifier, err := rawConfig.getTokenSource()
-	require.NoError(err)
-
-	require.NotNil(tokenSource)
-	require.Equal(sourceTypeServicePrincipal, tokenSourceType)
-	require.Equal("client-id", sourceIdentifier)
 }
